requirements.txt
alchemical==0.5.1
alembic==1.8.1
apifairy==1.3.0
apispec==6.0.2
click==8.1.3
Faker==15.3.2
Flask==2.2.2
Flask-HTTPAuth==4.7.0
flask-marshmallow==0.14.0
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.44
webargs==8.2.0
Werkzeug==2.2.2
